+++ backend/database.py
@@ -77,6 +77,7 @@ engine = create_engine(SQLALCHEMY_DATABASE_URL)
 
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
+
 Base = declarative_base()
 
 def get_db() -> Session:
@@ -105,52 +106,9 @@ def get_current_user(token: str, db: Session = Depends(get_db)):
         raise credentials_exception
     return user
 
-def insert_initial_data():
-    from models import Exercise
-    from exercise.mediapipe.exercise.waist import WaistExercise
-    from.exercise.mediapipe.exercise.squat import SquatExercise
-    from.exercise.mediapipe.exercise.leg import LegExercise
-    from.exercise.mediapipe.exercise.neck import NeckExercise
-    db = SessionLocal()
-    try:
-        exercises = [
-            Exercise(
-                name="neck",
-                count=5,
-                set=3,
-                time=0,
-                coordinate_list=[NeckExercise.get_neck_coordinates],
-            ),
-            Exercise(
-                name="hip",
-                count=5,
-                set=3,
-                time=0,
-                coordinate_list=[SquatExercise.get_squat_coordinates],
-            ),
-            Exercise(
-                name="leg",
-                count=5,
-                set=3,
-                time=10,
-                coordinate_list=[LegExercise.get_lunge_coordinates]
-            ),
-            Exercise(
-                name="waist",
-                count=10,
-                set=3,
-                time=0,
-                coordinate_list=WaistExercise.get_waist_coordinates(),
-            )
-        ]
-        db.add_all(exercises)
-        db.commit()
-    finally:
-        db.close()
 
 def init_db():
     Base.metadata.create_all(bind=engine)
-    insert_initial_data()
 
 if __name__ == "__main__":
     init_db()
